refactor: move build_context from chat to vectorstore
